+++ src/pages/home/components/LatestSection.tsx
@@ -30,7 +30,7 @@ export default function LastestSection({ rawConcertList, isLoading }: { rawConce
   }, [rawConcertList]);
 
   return (
-    <section className="mt-24 min-h-[100px] bg-neutral-100 lg:bg-white">
+    <section className="min-h-[100px] bg-neutral-100 lg:bg-white">
       {/* 手機板 */}
       <div className="lg:hidden">
         <MobileTitle title="最新活動" subtitle="Latest Events" subClass="text-neutral-200" />
@@ -47,8 +47,8 @@ export default function LastestSection({ rawConcertList, isLoading }: { rawConce
         )}
       </div>
       {/* 電腦板 */}
-      <div className="relative hidden h-[100vh] h-full lg:block">
-        <div className="relative mx-auto w-fit text-center text-4xl font-bold select-none">
+      <div className="relative mx-auto hidden h-full w-[90%] rounded-3xl lg:block lg:min-h-[400px] lg:overflow-hidden">
+        <div className="relative text-center text-4xl font-bold select-none">
           <h2 className="relative z-20 bg-gradient-to-r from-[#2D6ED0] to-[#2BC6CC] bg-clip-text text-[48px] text-transparent">最新活動</h2>
           <span className={`absolute top-[50%] left-1/2 z-10 w-[400px] -translate-x-1/2 text-center text-[40px] text-nowrap text-neutral-100`}>
             Latest Events